+++ packages/client/components/SpotlightResults.tsx
@@ -4,12 +4,12 @@ import React, {RefObject, useRef} from 'react'
 import {PreloadedQuery, usePreloadedQuery} from 'react-relay'
 import ReflectionGroup from './ReflectionGroup/ReflectionGroup'
 import {ElementHeight, ElementWidth} from '~/types/constEnums'
-import {SpotlightGroupsQuery} from '~/__generated__/SpotlightGroupsQuery.graphql'
 import useGroupMatrix from '../hooks/useGroupMatrix'
 import useResultsHeight from '~/hooks/useResultsHeight'
-import SpotlightGroupsEmptyState from './SpotlightGroupsEmptyState'
+import SpotlightResultsEmptyState from './SpotlightResultsEmptyState'
+import {SpotlightResultsQuery} from '~/__generated__/SpotlightResultsQuery.graphql'
 
-const SimilarGroups = styled('div')({
+const ResultsWrapper = styled('div')({
   padding: '40px 0px 24px',
   height: '100%',
   width: '100%',
@@ -35,14 +35,16 @@ const Column = styled('div')({
 
 interface Props {
   phaseRef: RefObject<HTMLDivElement>
-  queryRef: PreloadedQuery<SpotlightGroupsQuery>
+  queryRef: PreloadedQuery<SpotlightResultsQuery>
+  isSpotlightEntering: boolean
 }
 
-const SpotlightGroups = (props: Props) => {
-  const {phaseRef, queryRef} = props
-  const data = usePreloadedQuery<SpotlightGroupsQuery>(
+const SpotlightResults = (props: Props) => {
+  const {phaseRef, queryRef, isSpotlightEntering} = props
+
+  const data = usePreloadedQuery<SpotlightResultsQuery>(
     graphql`
-      query SpotlightGroupsQuery($reflectionGroupId: ID!, $searchQuery: String!, $meetingId: ID!) {
+      query SpotlightResultsQuery($reflectionGroupId: ID!, $searchQuery: String!, $meetingId: ID!) {
         viewer {
           similarReflectionGroups(
             reflectionGroupId: $reflectionGroupId
@@ -89,26 +91,30 @@ const SpotlightGroups = (props: Props) => {
   const groupMatrix = useGroupMatrix(similarReflectionGroups, resultsRef, phaseRef)
   const scrollHeight = useResultsHeight(resultsRef)
 
-  if (!similarReflectionGroups.length) return <SpotlightGroupsEmptyState />
   return (
-    <SimilarGroups>
-      <Scrollbar height={scrollHeight} ref={resultsRef}>
-        {groupMatrix.map((row) => (
-          <Column key={`row-${row[0]?.id}`}>
-            {row.map((group) => (
-              <ReflectionGroup
-                key={group.id}
-                meetingRef={meeting!}
-                phaseRef={phaseRef}
-                reflectionGroupRef={group}
-                expandedReflectionGroupPortalParentId='spotlight'
-              />
-            ))}
-          </Column>
-        ))}
-      </Scrollbar>
-    </SimilarGroups>
+    <ResultsWrapper>
+      {!similarReflectionGroups.length ? (
+        <SpotlightResultsEmptyState height={scrollHeight} />
+      ) : (
+        <Scrollbar height={scrollHeight} ref={resultsRef}>
+          {groupMatrix.map((row) => (
+            <Column key={`row-${row[0]?.id}`}>
+              {row.map((group) => (
+                <ReflectionGroup
+                  key={group.id}
+                  meetingRef={meeting!}
+                  phaseRef={phaseRef}
+                  reflectionGroupRef={group}
+                  expandedReflectionGroupPortalParentId='spotlight'
+                  isSpotlightEntering={isSpotlightEntering}
+                />
+              ))}
+            </Column>
+          ))}
+        </Scrollbar>
+      )}
+    </ResultsWrapper>
   )
 }
 
-export default SpotlightGroups
+export default SpotlightResults