+++ packages/client/components/TaskEditor/decorators.ts
@@ -1,22 +1,30 @@
 import EditorLink from './EditorLink'
-import {CompositeDecorator, EditorState} from 'draft-js'
+import {CompositeDecorator, ContentBlock, ContentState, EditorState} from 'draft-js'
 import Hashtag from './Hashtag'
 import Mention from './Mention'
+import SearchHighlight from './SearchHighlight'
 import TruncatedEllipsis from './TruncatedEllipsis'
 import {SetEditorState} from '../../types/draft'
 
-const findEntity = (entityType) => (contentBlock, callback, contentState) => {
-  contentBlock.findEntityRanges((character) => {
-    const entityKey = character.getEntity()
-    return entityKey !== null && contentState.getEntity(entityKey).getType() === entityType
-  }, callback)
-}
+const findEntity =
+  (entityType: string) =>
+  (
+    contentBlock: ContentBlock,
+    callback: (start: number, end: number) => void,
+    contentState: ContentState
+  ) => {
+    contentBlock.findEntityRanges((character) => {
+      const entityKey = character.getEntity()
+      return entityKey !== null && contentState.getEntity(entityKey).getType() === entityType
+    }, callback)
+  }
 
 const decorators = (
   getEditorState: () => EditorState | undefined,
-  setEditorState?: SetEditorState
-) =>
-  new CompositeDecorator([
+  setEditorState?: SetEditorState,
+  searchQuery?: string | null
+) => {
+  const compositeDecorator = [
     {
       strategy: findEntity('LINK'),
       component: EditorLink(getEditorState)
@@ -33,6 +41,39 @@ const decorators = (
       strategy: findEntity('TRUNCATED_ELLIPSIS'),
       component: TruncatedEllipsis(setEditorState)
     }
-  ])
+  ]
+
+  if (searchQuery) {
+    const findMatchingText = (
+      contentBlock: ContentBlock,
+      callback: (startIdx: number, endIdx: number) => void
+    ) => {
+      if (!searchQuery) {
+        return
+      }
+      const textLower = contentBlock.getText().toLowerCase()
+      const searchQueryLower = searchQuery.toLowerCase()
+
+      let start = 0
+      let foundAll = false
+      while (!foundAll) {
+        const index = textLower.indexOf(searchQueryLower, start)
+        if (index === -1) {
+          foundAll = true
+        } else {
+          start = index + 1
+          callback(index, index + searchQueryLower.length)
+        }
+      }
+    }
+
+    compositeDecorator.push({
+      strategy: findMatchingText,
+      component: SearchHighlight
+    })
+  }
+
+  return new CompositeDecorator(compositeDecorator)
+}
 
 export default decorators