+++ packages/client/hooks/useGroupMatrix.ts
@@ -1,11 +1,11 @@
-import {SpotlightGroupsQueryResponse} from './../__generated__/SpotlightGroupsQuery.graphql'
+import {SpotlightResultsQueryResponse} from './../__generated__/SpotlightResultsQuery.graphql'
 import {RefObject, useEffect, useState} from 'react'
 import {ElementWidth} from '../types/constEnums'
 import useResizeObserver from './useResizeObserver'
 import {MAX_SPOTLIGHT_COLUMNS} from '~/utils/constants'
 import useInitialRender from './useInitialRender'
 
-type Group = SpotlightGroupsQueryResponse['viewer']['similarReflectionGroups'][0]
+type Group = SpotlightResultsQueryResponse['viewer']['similarReflectionGroups'][0]
 
 const useGroupMatrix = (
   resultsGroups: readonly Group[],
@@ -13,6 +13,7 @@ const useGroupMatrix = (
   phaseRef: RefObject<HTMLDivElement>
 ) => {
   const isInit = useInitialRender()
+  const [groupMatrix, setGroupMatrix] = useState<Group[][]>([])
 
   const getColumnsCount = () => {
     const {current: el} = resultsRef
@@ -28,20 +29,18 @@ const useGroupMatrix = (
     return colCount
   }
 
-  const initMatrix = (): Group[][] => {
+  const initMatrix = () => {
     const columnsCount = getColumnsCount()
-    if (columnsCount === null) return []
+    if (columnsCount === null) return
     const matrix = Array.from(new Array(columnsCount), () => []) as Group[][]
     resultsGroups.forEach((group, idx) => {
       const columnIdx = idx % columnsCount
       matrix[columnIdx].push(group)
     })
-    return matrix
+    setGroupMatrix(matrix)
   }
 
-  const [groupMatrix, setGroupMatrix] = useState(initMatrix)
-
-  const getEmptiestColumnIdx = () => {
+  const getEmptiestColumnIdx = (groupMatrix) => {
     const columnsCount = getColumnsCount()
     // Use initColumns to get emptiest column vs emptiest column with a group in it.
     // For example, column 1 & column 2 both contain 1 group, group in column 2 is
@@ -57,40 +56,38 @@ const useGroupMatrix = (
 
   const updateMatrix = () => {
     const matrixGroupsIds = groupMatrix.flatMap((row) => row.map(({id}) => id))
-    const newGroup = resultsGroups.find((group) => !matrixGroupsIds.includes(group.id))
-    const refGroupsIds = resultsGroups.map(({id}) => id)
-    const removedGroupId = matrixGroupsIds.find((id) => !refGroupsIds.includes(id))
-    if (newGroup && removedGroupId) {
-      // added to kanban group. Remove old Spotlight group and add kanban group in place
-      const newMatrix = groupMatrix.map((row) =>
-        row.map((group) => (group.id === removedGroupId ? newGroup : group))
-      )
-      setGroupMatrix(newMatrix)
-    } else if (newGroup) {
-      // ungrouping added a new group
-      const emptiestColumnIdx = getEmptiestColumnIdx()
-      const matrixColumns = groupMatrix.length - 1
-      const newMatrix =
-        emptiestColumnIdx > matrixColumns // add to empty column
-          ? [...groupMatrix, [newGroup]]
-          : groupMatrix.map((row, idx) => (idx === emptiestColumnIdx ? [...row, newGroup] : row))
-      setGroupMatrix(newMatrix)
-    } else if (removedGroupId) {
-      // group added to another Spotlight group. Remove the old group & remove column if empty
-      const newMatrix = groupMatrix
-        .map((row) => row.filter((group) => group.id !== removedGroupId))
-        .filter(({length}) => length)
+    const newGroups = resultsGroups.filter(({id}) => !matrixGroupsIds.includes(id))
+    const resultsGroupsIds = resultsGroups.map(({id}) => id)
+    const removedGroupsIds = matrixGroupsIds.filter((id) => !resultsGroupsIds.includes(id))
+    if (!removedGroupsIds.length && !newGroups.length) return
+    // grouping results or updating the search query can remove groups
+    const filteredMatrix = groupMatrix
+      .map((row) => row.filter(({id}) => !removedGroupsIds.includes(id)))
+      .filter(({length}) => length) // remove column if empty
+    if (!newGroups.length) {
+      setGroupMatrix(filteredMatrix)
+    } else if (newGroups.length) {
+      // ungrouping results or updating the search query can add new groups
+      const newMatrix = newGroups.reduce((matrix, newGroup) => {
+        const emptiestColumnIdx = getEmptiestColumnIdx(matrix)
+        const matrixColumns = matrix.length - 1
+        return emptiestColumnIdx > matrixColumns
+          ? [...matrix, [newGroup]] // add to empty column
+          : matrix.map((row, idx) => (idx === emptiestColumnIdx ? [...row, newGroup] : row))
+      }, filteredMatrix)
       setGroupMatrix(newMatrix)
     }
   }
 
   useEffect(() => {
-    if (isInit) return
+    if (isInit) {
+      initMatrix()
+      return
+    }
     updateMatrix()
   }, [resultsGroups])
 
-  const createMatrix = () => setGroupMatrix(initMatrix())
-  useResizeObserver(createMatrix, phaseRef)
+  useResizeObserver(initMatrix, phaseRef)
 
   return groupMatrix
 }