+++ packages/client/hooks/useResultsHeight.ts
@@ -1,7 +1,7 @@
 import {RefObject, useLayoutEffect, useState} from 'react'
 import useResizeObserver from './useResizeObserver'
 
-// if results are remotely ungrouped, SpotlightGroups increases in height.
+// if results are remotely ungrouped, SpotlightResults increases in height.
 // to prevent the modal height from changing, use initial groups height
 const useResultsHeight = (resultsRef: RefObject<HTMLDivElement>) => {
   const [height, setHeight] = useState<number | string>('100%')
@@ -14,8 +14,10 @@ const useResultsHeight = (resultsRef: RefObject<HTMLDivElement>) => {
   }, [height])
 
   useResizeObserver(() => {
-    // when resized, set height to 100% so that useLayoutEffect can calc the new height
-    setHeight('100%')
+    // when resized & has ref, set height to 100% so that useLayoutEffect can calc the new height
+    if (resultsRef.current) {
+      setHeight('100%')
+    }
   }, resultsRef)
   return height
 }