+++ packages/server/graphql/mutations/endSprintPoker.ts
@@ -15,7 +15,8 @@ import {GQLContext} from '../graphql'
 import EndSprintPokerPayload from '../types/EndSprintPokerPayload'
 import sendMeetingEndToSegment from './helpers/endMeeting/sendMeetingEndToSegment'
 import sendNewMeetingSummary from './helpers/endMeeting/sendNewMeetingSummary'
-import {endSlackMeeting} from './helpers/notifySlack'
+import {endSlackMeeting} from './helpers/notifications/notifySlack'
+import {endMattermostMeeting} from './helpers/notifications/notifyMattermost'
 import removeEmptyTasks from './helpers/removeEmptyTasks'
 import getPhase from '../../utils/getPhase'
 
@@ -74,24 +75,24 @@ export default {
     stage.isComplete = true
     stage.endAt = now
 
-    const completedMeeting = ((await r
+    const completedMeeting = (await r
       .table('NewMeeting')
       .get(meetingId)
       .update(
         {
           endedAt: now,
           phases,
-          commentCount: (r
+          commentCount: r
             .table('Comment')
             .getAll(r.args(discussionIds), {index: 'discussionId'})
             .count()
-            .default(0) as unknown) as number,
+            .default(0) as unknown as number,
           storyCount
         },
         {returnChanges: true, nonAtomic: true}
       )('changes')(0)('new_val')
       .default(null)
-      .run()) as unknown) as MeetingPoker
+      .run()) as unknown as MeetingPoker
     if (!completedMeeting) {
       return standardError(new Error('Completed poker meeting does not exist'), {
         userId: viewerId
@@ -106,6 +107,7 @@ export default {
       // technically, this template could have mutated while the meeting was going on. but in practice, probably not
       dataLoader.get('meetingTemplates').load(templateId)
     ])
+    endMattermostMeeting(meetingId, teamId, dataLoader).catch(console.log)
     endSlackMeeting(meetingId, teamId, dataLoader).catch(console.log)
     sendMeetingEndToSegment(completedMeeting, meetingMembers as MeetingMember[], template)
     const isKill = phase && phase.phaseType !== 'ESTIMATE'
@@ -121,10 +123,7 @@ export default {
           meetingId
         })
     )
-    await r
-      .table('TimelineEvent')
-      .insert(events)
-      .run()
+    await r.table('TimelineEvent').insert(events).run()
 
     const data = {
       meetingId,