+++ packages/server/graphql/mutations/verifyEmail.ts
@@ -4,14 +4,13 @@ import getRethink from '../../database/rethinkDriver'
 import AuthIdentityLocal from '../../database/types/AuthIdentityLocal'
 import AuthToken from '../../database/types/AuthToken'
 import EmailVerification from '../../database/types/EmailVerification'
-import db from '../../db'
+import {getUserByEmail} from '../../postgres/queries/getUsersByEmails'
+import updateUser from '../../postgres/queries/updateUser'
 import createNewLocalUser from '../../utils/createNewLocalUser'
 import encodeAuthToken from '../../utils/encodeAuthToken'
 import rateLimit from '../rateLimit'
 import VerifyEmailPayload from '../types/VerifyEmailPayload'
 import bootstrapNewUser from './helpers/bootstrapNewUser'
-import updateUser from '../../postgres/queries/updateUser'
-import {getUserByEmail} from '../../postgres/queries/getUsersByEmails'
 
 export default {
   type: new GraphQLNonNull(VerifyEmailPayload),
@@ -56,16 +55,12 @@ export default {
       if (!localIdentity.isEmailVerified) {
         // mutative
         localIdentity.isEmailVerified = true
-        await Promise.all([
-          updateUser(
-            {
-              identities,
-              updatedAt: now
-            },
-            userId
-          ),
-          db.write('User', userId, {identities, updatedAt: now})
-        ])
+        await updateUser(
+          {
+            identities
+          },
+          userId
+        )
       }
       return {authToken, userId}
     }