+++ packages/server/graphql/types/EstimateStage.ts
@@ -13,9 +13,9 @@ import {SprintPokerDefaults} from '../../../client/types/constEnums'
 import EstimateStageDB from '../../database/types/EstimateStage'
 import {NewMeetingPhaseTypeEnum} from '../../database/types/GenericMeetingPhase'
 import MeetingPoker from '../../database/types/MeetingPoker'
-import db from '../../db'
 import getRedis from '../../utils/getRedis'
 import {GQLContext} from '../graphql'
+import isValid from '../isValid'
 import DiscussionThreadStage, {discussionThreadStageFields} from './DiscussionThreadStage'
 import EstimateUserScore from './EstimateUserScore'
 import NewMeetingStage, {newMeetingStageFields} from './NewMeetingStage'
@@ -62,7 +62,7 @@ const EstimateStage = new GraphQLObjectType<Source, GQLContext>({
         const {service} = integration
         const getDimensionName = async (meetingId: string) => {
           const meeting = await dataLoader.get('newMeetings').load(meetingId)
-          const {templateRefId} = meeting
+          const {templateRefId} = meeting as MeetingPoker
           const templateRef = await dataLoader.get('templateRefs').load(templateRefId)
           const {dimensions} = templateRef
           const dimensionRef = dimensions[dimensionRefIdx]
@@ -146,7 +146,7 @@ const EstimateStage = new GraphQLObjectType<Source, GQLContext>({
           dataLoader.get('newMeetings').load(meetingId),
           dataLoader.get('meetingTaskEstimates').load({taskId, meetingId})
         ])
-        const {templateRefId} = meeting
+        const {templateRefId} = meeting as MeetingPoker
         const templateRef = await dataLoader.get('templateRefs').load(templateRefId)
         const {dimensions} = templateRef
         const dimensionRef = dimensions[dimensionRefIdx]
@@ -167,12 +167,11 @@ const EstimateStage = new GraphQLObjectType<Source, GQLContext>({
     hoveringUsers: {
       type: new GraphQLNonNull(new GraphQLList(new GraphQLNonNull(User))),
       description: 'the users of the team members hovering the deck',
-      resolve: async ({id: stageId}) => {
+      resolve: async ({id: stageId}, _args: unknown, {dataLoader}) => {
         const redis = getRedis()
         const userIds = await redis.smembers(`pokerHover:${stageId}`)
         if (userIds.length === 0) return []
-        const users = await db.readMany('User', userIds)
-        return users
+        return (await dataLoader.get('users').loadMany(userIds)).filter(isValid)
       }
     },
     scores: {