+++ packages/server/utils/getRedis.ts
@@ -3,7 +3,7 @@ import Redis from 'ioredis'
 let redis: Redis.Redis
 const getRedis = () => {
   if (!redis) {
-    redis = new Redis(process.env.REDIS_URL)
+    redis = new Redis(process.env.REDIS_URL, {connectionName: 'getRedis'})
   }
   return redis
 }