+++ CHANGELOG.md
@@ -5,6 +5,43 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 This CHANGELOG follows conventions [outlined here](http://keepachangelog.com/).
 
+## 6.40.0 2021-Dec-09
+
+### Added
+
+- Add mattermost notification integration (#5550)
+- Add datadog tracing to compiled graphql queries (#5681)
+- Add hardDeleteUser mutation (#5475)
+- Add spotlightSearchQuery to similarReflectionGroups queries (#5735)
+- Spotlight — highlight search keyword (#5716)
+- Spotlight feature flag (#5699)
+- Add calculated domain field to User (#5719)
+
+### Changed
+
+- Extend GQL Timeout for GraphiQL Queries (#5577)
+
+### Fixed
+
+- Prevent ungrouping Spotlight results (#5704)
+- Fix spotlight not loading similar reflections (#5728)
+- Mitigate safari popups on SSO (#5617)
+- Fix more implicit any errors (#5740)
+
+## 6.39.0 2021-Dec-02
+
+### Added
+
+- Language tooltip is updated on New Meeting CheckIn Greeting (#5682)
+- Spotlight: support searchQuery business logic for similarReflections (#5469)
+- Spotlight: animate remote reflection when it enters or leaves Spotlight (#5600)
+
+### Fixed
+
+- Fix remotely ungrouped results position in Spotlight (#5708)
+- Recently used emojis are managed using local storage (#5656)
+- Returning refreshed token from fresh atlassian auth loader (#5729)
+
 ## 6.38.0 2021-Nov-24
 
 ### Added
@@ -23,6 +60,7 @@ This CHANGELOG follows conventions [outlined here](http://keepachangelog.com/).
 - Fix dail_user_stat to order by user counts (#5693)
 - remove hasAuth from scopephaseareaadd (#5711)
 - Fix no implicit any errors (Part 3 of 8) (#5697)
+- Invitation required screen right after creating a new team (#5721)
 
 ### Changed
 
@@ -31,6 +69,7 @@ This CHANGELOG follows conventions [outlined here](http://keepachangelog.com/).
 - bump eslint to v8 (#5695)
 - favor built assets over \_\_STATIC_IMAGES\_\_ (#4847)
 - Remove wrtc signal server (#5698)
+- support eslint v7+ (#5709)
 
 ## 6.37.0 2021-Nov-18